Rename extensions classes to avoid conflicts in used directly
