+++ src/Tingle.EventBus.Transports.Amazon.Kinesis/AmazonKinesisEventBusBuilderExtensions.cs
@@ -6,7 +6,7 @@ namespace Microsoft.Extensions.DependencyInjection;
 /// <summary>
 /// Extension methods on <see cref="EventBusBuilder"/> for Amazon Kinesis.
 /// </summary>
-public static class EventBusBuilderExtensions
+public static class AmazonKinesisEventBusBuilderExtensions
 {
     /// <summary>Add Amazon Kinesis transport.</summary>
     /// <param name="builder">The <see cref="EventBusBuilder"/> to add to.</param>
