+++ src/Tingle.EventBus.Transports.Amazon.Sqs/AmazonSqsEventBusBuilderExtensions.cs
@@ -6,7 +6,7 @@ namespace Microsoft.Extensions.DependencyInjection;
 /// <summary>
 /// Extension methods on <see cref="EventBusBuilder"/> for Amazon SQS.
 /// </summary>
-public static class EventBusBuilderExtensions
+public static class AmazonSqsEventBusBuilderExtensions
 {
     /// <summary>Add Amazon SQS transport.</summary>
     /// <param name="builder">The <see cref="EventBusBuilder"/> to add to.</param>
